add the font from unlaunch! (part 1)
Former-commit-id: 5f2040c6f2b414f74d587a57603d16f94eeca77a
